Refactors ViewTeams
